refactor: inline publication string rules
