+++ pkg/rules/string_test.go
@@ -124,7 +124,6 @@ var stringDNSLabelTestCases = []*struct {
 	in         string
 	shouldFail bool
 }{
-	// cspell:disable
 	{"test", false},
 	{"s", false},
 	{"test-this", false},
@@ -139,7 +138,6 @@ var stringDNSLabelTestCases = []*struct {
 	{"test this", true},
 	{"1_2", true},
 	{"LOL", true},
-	// cspell:enable
 }
 
 func TestStringDNSLabel(t *testing.T) {
@@ -167,7 +165,6 @@ var stringDNSSubdomainTestCases = []*struct {
 	in         string
 	shouldFail bool
 }{
-	// cspell:disable
 	{"s", false},
 	{"sa", false},
 	{"a-1", false},
@@ -203,7 +200,6 @@ var stringDNSSubdomainTestCases = []*struct {
 	{".2.3.4", true},
 	{"1a.2B.3c.4d", true},
 	{"a--b--c.", true},
-	// cspell:enable
 }
 
 func TestStringDNSSubdomain(t *testing.T) {
@@ -231,7 +227,6 @@ var stringASCIITestCases = []*struct {
 	in         string
 	shouldFail bool
 }{
-	// cspell:disable
 	{"", false},
 	{"foobar", false},
 	{"0987654321", false},
@@ -243,7 +238,6 @@ var stringASCIITestCases = []*struct {
 	{"ｘｙｚ０９８", true},
 	{"１２３456", true},
 	{"ｶﾀｶﾅ", true},
-	// cspell:enable
 }
 
 func TestStringASCII(t *testing.T) {
@@ -271,7 +265,6 @@ var stringUUIDTestCases = []*struct {
 	in         string
 	shouldFail bool
 }{
-	// cspell:disable
 	{"00000000-0000-0000-0000-000000000000", false},
 	{"e190c630-8873-11ee-b9d1-0242ac120002", false},
 	{"79258D24-01A7-47E5-ACBB-7E762DE52298", false},
@@ -287,7 +280,6 @@ var stringUUIDTestCases = []*struct {
 	{"934859", true},
 	{"987fbc9-4bed-3078-cf07a-9141ba07c9F3", true},
 	{"aaaaaaaa-1111-1111-aaaG-111111111111", true},
-	// cspell:enable
 }
 
 func TestStringUUID(t *testing.T) {
@@ -315,7 +307,6 @@ var stringEmailTestCases = []*struct {
 	in         string
 	shouldFail bool
 }{
-	// cspell:disable
 	{"test@mail.com", false},
 	{"Dörte@Sörensen.example.com", false},
 	{"θσερ@εχαμπλε.ψομ", false},
@@ -332,7 +323,6 @@ var stringEmailTestCases = []*struct {
 	{"test@email.", true},
 	{"@email.com", true},
 	{`"@email.com`, true},
-	// cspell:enable
 }
 
 func TestStringEmail(t *testing.T) {
@@ -390,7 +380,6 @@ var stringMACTestCases = []*struct {
 	in         string
 	shouldFail bool
 }{
-	// cspell:disable
 	{"3D:F2:C9:A6:B3:4F", false},
 	{"00:25:96:FF:FE:12:34:56", false},
 	{"3D-F2-C9-A6-B3:4F", true},
@@ -398,7 +387,6 @@ var stringMACTestCases = []*struct {
 	{"", true},
 	{"abacaba", true},
 	{"0025:96FF:FE12:3456", true},
-	// cspell:enable
 }
 
 func TestStringMAC(t *testing.T) {
@@ -426,7 +414,6 @@ var stringIPTestCases = []*struct {
 	in         string
 	shouldFail bool
 }{
-	// cspell:disable
 	{"10.0.0.1", false},
 	{"172.16.0.1", false},
 	{"192.168.0.1", false},
@@ -438,7 +425,6 @@ var stringIPTestCases = []*struct {
 	{"", true},
 	{"172.16.256.255", true},
 	{"192.168.255.256", true},
-	// cspell:enable
 }
 
 func TestStringIP(t *testing.T) {
@@ -466,7 +452,6 @@ var stringIPv4TestCases = []*struct {
 	in         string
 	shouldFail bool
 }{
-	// cspell:disable
 	{"10.0.0.1", false},
 	{"172.16.0.1", false},
 	{"192.168.0.1", false},
@@ -477,7 +462,6 @@ var stringIPv4TestCases = []*struct {
 	{"2001:cdba:0000:0000:0000:0000:3257:9652", true},
 	{"2001:cdba:0:0:0:0:3257:9652", true},
 	{"2001:cdba::3257:9652", true},
-	// cspell:enable
 }
 
 func TestStringIPv4(t *testing.T) {
@@ -505,7 +489,6 @@ var stringIPv6TestCases = []*struct {
 	in         string
 	shouldFail bool
 }{
-	// cspell:disable
 	{"2001:cdba:0000:0000:0000:0000:3257:9652", false},
 	{"2001:cdba:0:0:0:0:3257:9652", false},
 	{"2001:cdba::3257:9652", false},
@@ -516,7 +499,6 @@ var stringIPv6TestCases = []*struct {
 	{"192.168.255.256", true},
 	{"172.16.255.254", true},
 	{"172.16.256.255", true},
-	// cspell:enable
 }
 
 func TestStringIPv6(t *testing.T) {
@@ -544,7 +526,6 @@ var stringCIDRTestCases = []*struct {
 	in         string
 	shouldFail bool
 }{
-	// cspell:disable
 	{"10.0.0.0/0", false},
 	{"10.0.0.1/8", false},
 	{"172.16.0.1/16", false},
@@ -558,7 +539,6 @@ var stringCIDRTestCases = []*struct {
 	{"192.168.255.256/24", true},
 	{"172.16.256.255/16", true},
 	{"2001:cdba:0000:0000:0000:0000:3257:9652/256", true},
-	// cspell:enable
 }
 
 func TestStringCIDR(t *testing.T) {
@@ -586,7 +566,6 @@ var stringCIDRv4TestCases = []*struct {
 	in         string
 	shouldFail bool
 }{
-	// cspell:disable
 	{"0.0.0.0/0", false},
 	{"10.0.0.0/8", false},
 	{"172.16.0.0/16", false},
@@ -607,7 +586,6 @@ var stringCIDRv4TestCases = []*struct {
 	{"2001:cdba:0:0:0:0:3257:9652/32", true},
 	{"2001:cdba::3257:9652/16", true},
 	{"172.56.1.0/16", true},
-	// cspell:enable
 }
 
 func TestStringCIDRv4(t *testing.T) {
@@ -635,7 +613,6 @@ var stringCIDRv6TestCases = []*struct {
 	in         string
 	shouldFail bool
 }{
-	// cspell:disable
 	{"2001:cdba:0000:0000:0000:0000:3257:9652/64", false},
 	{"2001:cdba:0:0:0:0:3257:9652/32", false},
 	{"2001:cdba::3257:9652/16", false},
@@ -649,7 +626,6 @@ var stringCIDRv6TestCases = []*struct {
 	{"172.16.255.254/16", true},
 	{"172.16.256.255/16", true},
 	{"2001:cdba:0000:0000:0000:0000:3257:9652/256", true},
-	// cspell:enable
 }
 
 func TestStringCIDRv6(t *testing.T) {
@@ -1193,7 +1169,6 @@ var stringTitleTestCases = []*struct {
 	in         string
 	shouldFail bool
 }{
-	// cspell:disable
 	{"", true},
 	{"a", true},
 	{"A", false},
@@ -1208,7 +1183,6 @@ var stringTitleTestCases = []*struct {
 	{"With_underscore", false},
 	{"unicode \xe2\x80\xa8 line separator", true},
 	{"Unicode \xe2\x80\xa8 Line Separator", false},
-	// cspell:enable
 }
 
 func TestStringTitle(t *testing.T) {
@@ -1508,7 +1482,6 @@ var stringMatchFileSystemPathTestCases = []*struct {
 	pattern, in string
 	shouldFail  bool
 }{
-	// cspell:disable
 	{"abc", "abc", false},
 	{"*", "abc", false},
 	{"*c", "abc", false},
@@ -1565,7 +1538,6 @@ var stringMatchFileSystemPathTestCases = []*struct {
 	{"a[", "ab", true},
 	{"a[", "x", true},
 	{"a/b[", "x", true},
-	// cspell:enable
 }
 
 func TestStringMatchFileSystemPath(t *testing.T) {
@@ -1597,7 +1569,6 @@ var stringRegexpTestCases = []*struct {
 	in         string
 	shouldFail bool
 }{
-	// cspell:disable
 	{``, false},
 	{`.`, false},
 	{`^.$`, false},
@@ -1627,7 +1598,6 @@ var stringRegexpTestCases = []*struct {
 	{`a*+`, true},
 	{`\x`, true},
 	{strings.Repeat(`\pL`, 27000), true},
-	// cspell:enable
 }
 
 func TestStringRegexp(t *testing.T) {
@@ -1882,7 +1852,6 @@ var stringAlphaTestCases = []*struct {
 	in         string
 	shouldFail bool
 }{
-	// cspell:disable
 	{"test", false},
 	{"tEsT", false},
 	{"s", false},
@@ -1893,7 +1862,6 @@ var stringAlphaTestCases = []*struct {
 	{" ", true},
 	{"test1", true},
 	{"tęst", true},
-	// cspell:enable
 }
 
 func TestStringAlpha(t *testing.T) {
@@ -1921,7 +1889,6 @@ var stringAlphanumericTestCases = []*struct {
 	in         string
 	shouldFail bool
 }{
-	// cspell:disable
 	{"test", false},
 	{"tEsT", false},
 	{"s", false},
@@ -1936,7 +1903,6 @@ var stringAlphanumericTestCases = []*struct {
 	{" ", true},
 	{"tęst", true},
 	{"tęst1", true},
-	// cspell:enable
 }
 
 func TestStringAlphanumeric(t *testing.T) {
@@ -1964,7 +1930,6 @@ var stringAlphaUnicodeTestCases = []*struct {
 	in         string
 	shouldFail bool
 }{
-	// cspell:disable
 	{"test", false},
 	{"tEsT", false},
 	{"s", false},
@@ -1979,7 +1944,6 @@ var stringAlphaUnicodeTestCases = []*struct {
 	{"test1", true},
 	{"汉语!", true},
 	{"1汉语", true},
-	// cspell:enable
 }
 
 func TestStringAlphaUnicode(t *testing.T) {
@@ -2007,7 +1971,6 @@ var stringAlphanumericUnicodeTestCases = []*struct {
 	in         string
 	shouldFail bool
 }{
-	// cspell:disable
 	{"test", false},
 	{"tEsT", false},
 	{"s", false},
@@ -2027,7 +1990,6 @@ var stringAlphanumericUnicodeTestCases = []*struct {
 	{"汉语!", true},
 	{"-921", true},
 	{" 1", true},
-	// cspell:enable
 }
 
 func TestStringAlphanumericUnicode(t *testing.T) {
@@ -2055,7 +2017,6 @@ var stringFQDNTestCases = []*struct {
 	in         string
 	shouldFail bool
 }{
-	// cspell:disable
 	{"test.example.com", false},
 	{"example.com", false},
 	{"example24.com", false},
@@ -2076,7 +2037,6 @@ var stringFQDNTestCases = []*struct {
 	{"2001:cdba:0:0:0:0:3257:9652", true},
 	{"2001:cdba::3257:9652", true},
 	{"", true},
-	// cspell:enable
 }
 
 func TestStringFQDN(t *testing.T) {
@@ -2118,7 +2078,6 @@ var stringK8sQualifiedNameTestCases = []*struct {
 	in          string
 	expectedErr error
 }{
-	// cspell:disable
 	{"simple", nil},
 	{"now-with-dashes", nil},
 	{"1-starts-with-num", nil},
@@ -2151,7 +2110,6 @@ var stringK8sQualifiedNameTestCases = []*struct {
 	{strings.Repeat("a", 64), errK8sQualifiedNameNamePartLength},
 	{strings.Repeat("a", 254) + "/abc", errK8sQualifiedNamePrefixLength},
 	{strings.Repeat("a", 253) + "/" + strings.Repeat("b", 64), errors.New("length must be between 1 and 317")},
-	// cspell:enable
 }
 
 func TestStringKubernetesQualifiedName(t *testing.T) {
@@ -2175,4 +2133,574 @@ func BenchmarkStringKubernetesQualifiedName(b *testing.B) {
 			_ = rule.Validate(tc.in)
 		}
 	}
+}
+
+var validISBNTestCases = map[string]string{
+	"isbn 10 hyphenated":                  "0-306-40615-2",
+	"isbn 10 plain":                       "0306406152",
+	"isbn 10 x check":                     "0-9752298-0-X",
+	"isbn 10 spaced":                      "0 9752298 0 x",
+	"isbn 10 library converter numeric":   "0394170660",
+	"isbn 10 library converter alternate": "0717941728",
+	"isbn 10 library converter x check":   "087779443X",
+	"isbn 10 MARC hyphenated":             "0-87068-693-3",
+	"isbn 13 hyphenated":                  "978-0-306-40615-7",
+	"isbn 13 plain":                       "9780306406157",
+	"isbn 13 grouped":                     "978-3-16-148410-0",
+	"isbn 13 agency manual hyphenated":    "978-92-95055-12-4",
+	"isbn 13 agency manual spaced":        "978 92 95055 12 4",
+	"isbn 13 agency manual compact":       "9789295055124",
+	"isbn 13 agency manual hardback":      "978-951-45-9693-3",
+	"isbn 13 agency manual paperback":     "978-951-45-9694-0",
+	"isbn 13 agency manual PDF":           "978-951-45-9695-7",
+	"isbn 13 agency manual EPUB":          "978-951-45-9696-4",
+	"isbn 13 library converter first":     "9780060723804",
+	"isbn 13 library converter second":    "9780060799748",
+	"isbn 13 979 prefix":                  "979-10-90636-07-1",
+}
+
+var invalidISBNTestCases = map[string]string{
+	"empty":                               "",
+	"isbn 10 failed check":                "0-306-40615-3",
+	"isbn 10 x check mutation":            "0877794430",
+	"isbn 10 x in body":                   "08777X443X",
+	"isbn 10 x in fourth position":        "087X79443X",
+	"isbn 10 short":                       "087779443",
+	"isbn 10 trailing space":              "087779443X ",
+	"isbn 13 failed check":                "978-0-306-40615-8",
+	"isbn 13 manual check mutation":       "978-92-95055-12-5",
+	"isbn 13 checksum valid wrong prefix": "9779295055125",
+	"isbn 13 x in body":                   "978-92-X5055-12-4",
+	"isbn 13 x check":                     "978-92-95055-12-X",
+	"isbn 13 en dash separators":          "978–92–95055–12–4",
+	"isbn 13 full width digits":           "９７８９２９５０５５１２４",
+	"isbn 13 display prefix":              "ISBN 978-92-95055-12-4",
+	"repeated separator":                  "978--0-306-40615-7",
+	"letters":                             "abc",
+}
+
+func TestStringISBN(t *testing.T) {
+	rule := StringISBN()
+	t.Run("valid inputs", func(t *testing.T) {
+		for name, input := range validISBNTestCases {
+			t.Run(name, func(t *testing.T) {
+				assert.NoError(t, rule.Validate(input))
+			})
+		}
+	})
+	t.Run("invalid inputs", func(t *testing.T) {
+		for name, input := range invalidISBNTestCases {
+			t.Run(name, func(t *testing.T) {
+				err := rule.Validate(input)
+				assert.EqualError(
+					t,
+					err,
+					"string must be a valid International Standard Book Number (ISBN) in ISBN-10 or ISBN-13 format",
+				)
+				assert.True(t, govy.HasErrorCode(err, ErrorCodeStringISBN))
+			})
+		}
+	})
+}
+
+func TestStringISBN_VeryLargeInvalid(t *testing.T) {
+	input := strings.Repeat("0", 1<<20)
+	tests := map[string]struct {
+		rule    govy.Rule[string]
+		message string
+		code    govy.ErrorCode
+	}{
+		"isbn": {
+			rule:    StringISBN(),
+			message: "string must be a valid International Standard Book Number (ISBN) in ISBN-10 or ISBN-13 format",
+			code:    ErrorCodeStringISBN,
+		},
+		"isbn-10": {
+			rule:    StringISBN10(),
+			message: "string must be a valid International Standard Book Number (ISBN) in ISBN-10 format",
+			code:    ErrorCodeStringISBN10,
+		},
+		"isbn-13": {
+			rule:    StringISBN13(),
+			message: "string must be a valid International Standard Book Number (ISBN) in ISBN-13 format",
+			code:    ErrorCodeStringISBN13,
+		},
+	}
+	for name, test := range tests {
+		t.Run(name, func(t *testing.T) {
+			err := test.rule.Validate(input)
+			assert.EqualError(t, err, test.message)
+			assert.True(t, govy.HasErrorCode(err, test.code))
+		})
+	}
+}
+
+func BenchmarkStringISBN(b *testing.B) {
+	benchmarkStringPublicationRule(
+		b,
+		StringISBN(),
+		validISBNTestCases,
+		invalidISBNTestCases,
+	)
+}
+
+func BenchmarkStringISBNVeryLargeInvalid(b *testing.B) {
+	rule := StringISBN()
+	input := strings.Repeat("0", 1<<20)
+
+	for b.Loop() {
+		_ = rule.Validate(input)
+	}
+	b.ReportMetric(1, "validations/op")
+}
+
+var validISBN10TestCases = map[string]string{
+	"hyphenated":                  "0-306-40615-2",
+	"plain":                       "0306406152",
+	"x check":                     "0-9752298-0-X",
+	"spaced":                      "0 9752298 0 x",
+	"library converter numeric":   "0394170660",
+	"library converter alternate": "0717941728",
+	"library converter x check":   "087779443X",
+	"MARC hyphenated":             "0-87068-693-3",
+}
+
+var invalidISBN10TestCases = map[string]string{
+	"empty":                "",
+	"failed check":         "0-306-40615-3",
+	"x check mutation":     "0877794430",
+	"x in body":            "08777X443X",
+	"x in fourth position": "087X79443X",
+	"short":                "087779443",
+	"trailing space":       "087779443X ",
+	"isbn 13":              "978-0-306-40615-7",
+	"isbn 13 plain":        "9780306406157",
+	"repeated separator":   "0-306--40615-2",
+}
+
+func TestStringISBN10(t *testing.T) {
+	rule := StringISBN10()
+	t.Run("valid inputs", func(t *testing.T) {
+		for name, input := range validISBN10TestCases {
+			t.Run(name, func(t *testing.T) {
+				assert.NoError(t, rule.Validate(input))
+			})
+		}
+	})
+	t.Run("invalid inputs", func(t *testing.T) {
+		for name, input := range invalidISBN10TestCases {
+			t.Run(name, func(t *testing.T) {
+				err := rule.Validate(input)
+				assert.EqualError(
+					t,
+					err,
+					"string must be a valid International Standard Book Number (ISBN) in ISBN-10 format",
+				)
+				assert.True(t, govy.HasErrorCode(err, ErrorCodeStringISBN10))
+			})
+		}
+	})
+}
+
+func BenchmarkStringISBN10(b *testing.B) {
+	benchmarkStringPublicationRule(
+		b,
+		StringISBN10(),
+		validISBN10TestCases,
+		invalidISBN10TestCases,
+	)
+}
+
+var validISBN13TestCases = map[string]string{
+	"hyphenated":               "978-0-306-40615-7",
+	"plain":                    "9780306406157",
+	"grouped":                  "978-3-16-148410-0",
+	"agency manual hyphenated": "978-92-95055-12-4",
+	"agency manual spaced":     "978 92 95055 12 4",
+	"agency manual compact":    "9789295055124",
+	"agency manual hardback":   "978-951-45-9693-3",
+	"agency manual paperback":  "978-951-45-9694-0",
+	"agency manual PDF":        "978-951-45-9695-7",
+	"agency manual EPUB":       "978-951-45-9696-4",
+	"library converter first":  "9780060723804",
+	"library converter second": "9780060799748",
+	"979 prefix":               "979-10-90636-07-1",
+}
+
+var invalidISBN13TestCases = map[string]string{
+	"empty":                       "",
+	"isbn 10":                     "0-306-40615-2",
+	"failed check":                "978-0-306-40615-8",
+	"manual check mutation":       "978-92-95055-12-5",
+	"checksum valid wrong prefix": "9779295055125",
+	"x in body":                   "978-92-X5055-12-4",
+	"x check":                     "978-92-95055-12-X",
+	"en dash separators":          "978–92–95055–12–4",
+	"full width digits":           "９７８９２９５０５５１２４",
+	"display prefix":              "ISBN 978-92-95055-12-4",
+	"invalid prefix":              "9770306406157",
+	"trailing space":              "978 0 306 40615 7 ",
+}
+
+func TestStringISBN13(t *testing.T) {
+	rule := StringISBN13()
+	t.Run("valid inputs", func(t *testing.T) {
+		for name, input := range validISBN13TestCases {
+			t.Run(name, func(t *testing.T) {
+				assert.NoError(t, rule.Validate(input))
+			})
+		}
+	})
+	t.Run("invalid inputs", func(t *testing.T) {
+		for name, input := range invalidISBN13TestCases {
+			t.Run(name, func(t *testing.T) {
+				err := rule.Validate(input)
+				assert.EqualError(
+					t,
+					err,
+					"string must be a valid International Standard Book Number (ISBN) in ISBN-13 format",
+				)
+				assert.True(t, govy.HasErrorCode(err, ErrorCodeStringISBN13))
+			})
+		}
+	})
+}
+
+func TestISBNPredicatesMatchReference(t *testing.T) {
+	tests := map[string]struct {
+		predicate func(string) bool
+		reference func(string) bool
+		inputs    []map[string]string
+	}{
+		"isbn": {
+			predicate: isISBN,
+			reference: referenceISBN,
+			inputs: []map[string]string{
+				validISBNTestCases,
+				invalidISBNTestCases,
+			},
+		},
+		"isbn-10": {
+			predicate: isISBN10,
+			reference: referenceISBN10,
+			inputs: []map[string]string{
+				validISBN10TestCases,
+				invalidISBN10TestCases,
+			},
+		},
+		"isbn-13": {
+			predicate: isISBN13,
+			reference: referenceISBN13,
+			inputs: []map[string]string{
+				validISBN13TestCases,
+				invalidISBN13TestCases,
+			},
+		},
+	}
+	for name, test := range tests {
+		t.Run(name, func(t *testing.T) {
+			testStringPredicateMatchesReference(t, test.predicate, test.reference, test.inputs...)
+		})
+	}
+}
+
+func BenchmarkStringISBN13(b *testing.B) {
+	benchmarkStringPublicationRule(
+		b,
+		StringISBN13(),
+		validISBN13TestCases,
+		invalidISBN13TestCases,
+	)
+}
+
+var validISSNTestCases = map[string]string{
+	"numeric check":       "2049-3630",
+	"numeric example":     "0378-5955",
+	"uppercase x":         "2434-561X",
+	"lowercase x":         "2434-561x",
+	"manual numeric":      "1106-1111",
+	"manual uppercase x":  "1092-003X",
+	"library check digit": "0317-8471",
+	"numeric 2162":        "2162-3546",
+	"numeric 1548":        "1548-7180",
+	"uppercase x 1204":    "1204-539X",
+}
+
+// invalidISSNTestCases includes exact compact construction examples from the
+// [ISSN Manual, May 2025] because StringISSN requires the ASCII-hyphenated form.
+// Their derived hyphenated forms are accepted in validISSNTestCases.
+//
+// [ISSN Manual, May 2025]: https://www.issn.org/wp-content/uploads/2025/05/Manual-ISSN_ENG-marc21_May2025.pdf
+var invalidISSNTestCases = map[string]string{
+	"empty":                      "",
+	"missing hyphen":             "20493630",
+	"manual compact 2162":        "21623546",
+	"manual compact 1548":        "15487180",
+	"failed check":               "2049-3631",
+	"numeric check mutation":     "1106-1112",
+	"uppercase x check mutation": "1092-0030",
+	"wrong grouping":             "204-93630",
+	"x before check":             "2049-36X0",
+	"hyphen as check":            "2049-363-",
+	"unicode hyphen":             "1106–1111",
+	"U+2010 hyphen":              "1092‐003X",
+	"space separator":            "1106 1111",
+	"display prefix":             "ISSN 1106-1111",
+	"trailing newline":           "1106-1111\n",
+	"full width digits":          "１１０６-１１１１",
+}
+
+func TestStringISSN(t *testing.T) {
+	rule := StringISSN()
+	t.Run("valid inputs", func(t *testing.T) {
+		for name, input := range validISSNTestCases {
+			t.Run(name, func(t *testing.T) {
+				assert.NoError(t, rule.Validate(input))
+			})
+		}
+	})
+	t.Run("invalid inputs", func(t *testing.T) {
+		for name, input := range invalidISSNTestCases {
+			t.Run(name, func(t *testing.T) {
+				err := rule.Validate(input)
+				assert.EqualError(t, err, "string must be a valid International Standard Serial Number (ISSN)")
+				assert.True(t, govy.HasErrorCode(err, ErrorCodeStringISSN))
+			})
+		}
+	})
+}
+
+func TestISSNPredicateMatchesReference(t *testing.T) {
+	format := regexp.MustCompile(`^\d{4}-\d{3}[0-9Xx]$`)
+	testStringPredicateMatchesReference(
+		t,
+		isISSN,
+		func(s string) bool {
+			return referenceISSN(format, s)
+		},
+		validISSNTestCases,
+		invalidISSNTestCases,
+	)
+}
+
+func BenchmarkStringISSN(b *testing.B) {
+	benchmarkStringPublicationRule(
+		b,
+		StringISSN(),
+		validISSNTestCases,
+		invalidISSNTestCases,
+	)
+}
+
+func Benchmark_isISSN(b *testing.B) {
+	for b.Loop() {
+		for name, input := range validISSNTestCases {
+			if !isISSN(input) {
+				b.Fatalf("%s: expected valid ISSN", name)
+			}
+		}
+		for name, input := range invalidISSNTestCases {
+			if isISSN(input) {
+				b.Fatalf("%s: expected invalid ISSN", name)
+			}
+		}
+	}
+	b.ReportMetric(float64(len(validISSNTestCases)+len(invalidISSNTestCases)), "validations/op")
+}
+
+func benchmarkStringPublicationRule(
+	b *testing.B,
+	rule govy.Rule[string],
+	validInputs map[string]string,
+	invalidInputs map[string]string,
+) {
+	b.Helper()
+	for b.Loop() {
+		for _, in := range validInputs {
+			_ = rule.Validate(in)
+		}
+		for _, in := range invalidInputs {
+			_ = rule.Validate(in)
+		}
+	}
+	b.ReportMetric(float64(len(validInputs)+len(invalidInputs)), "validations/op")
+}
+
+type stringPredicate func(string) bool
+
+func testStringPredicateMatchesReference(
+	t *testing.T,
+	predicate stringPredicate,
+	reference stringPredicate,
+	inputs ...map[string]string,
+) {
+	t.Helper()
+	for _, inputSet := range inputs {
+		for name, input := range inputSet {
+			assertStringPredicateMatchesReference(t, predicate, reference, name, input)
+			testStringPredicateByteEdits(t, predicate, reference, input)
+		}
+	}
+}
+
+func testStringPredicateByteEdits(
+	t *testing.T,
+	predicate stringPredicate,
+	reference stringPredicate,
+	input string,
+) {
+	t.Helper()
+	for position := range len(input) {
+		candidate := input[:position] + input[position+1:]
+		assertStringPredicateMatchesReference(t, predicate, reference, "byte deletion", candidate)
+	}
+
+	replacement := []byte(input)
+	for position := range len(replacement) {
+		original := replacement[position]
+		for value := range 256 {
+			replacement[position] = byte(value)
+			assertStringPredicateMatchesReference(
+				t,
+				predicate,
+				reference,
+				"byte replacement",
+				string(replacement),
+			)
+		}
+		replacement[position] = original
+	}
+
+	for position := range len(input) + 1 {
+		insertion := make([]byte, len(input)+1)
+		copy(insertion, input[:position])
+		copy(insertion[position+1:], input[position:])
+		for value := range 256 {
+			insertion[position] = byte(value)
+			assertStringPredicateMatchesReference(
+				t,
+				predicate,
+				reference,
+				"byte insertion",
+				string(insertion),
+			)
+		}
+	}
+}
+
+func assertStringPredicateMatchesReference(
+	t *testing.T,
+	predicate stringPredicate,
+	reference stringPredicate,
+	name string,
+	input string,
+) {
+	t.Helper()
+	got := predicate(input)
+	expected := reference(input)
+	if got != expected {
+		t.Fatalf("%s for %q: got %t, expected %t", name, input, got, expected)
+	}
+}
+
+func referenceISBN(s string) bool {
+	return referenceISBN10(s) || referenceISBN13(s)
+}
+
+func referenceISBN10(s string) bool {
+	isbn, ok := referenceNormalizeISBN(s)
+	if !ok || len(isbn) != isbn10Length {
+		return false
+	}
+
+	sum := 0
+	for i := range isbn10Length - 1 {
+		if !isASCIIDigit(isbn[i]) {
+			return false
+		}
+		sum += int(isbn[i]-'0') * (isbn10Length - i)
+	}
+
+	switch checkDigit := isbn[isbn10Length-1]; {
+	case isASCIIDigit(checkDigit):
+		sum += int(checkDigit - '0')
+	case checkDigit == 'X' || checkDigit == 'x':
+		sum += 10
+	default:
+		return false
+	}
+	return sum%11 == 0
+}
+
+func referenceISBN13(s string) bool {
+	isbn, ok := referenceNormalizeISBN(s)
+	if !ok || len(isbn) != isbn13Length ||
+		(!strings.HasPrefix(isbn, "978") && !strings.HasPrefix(isbn, "979")) {
+		return false
+	}
+
+	sum := 0
+	for i := range isbn13Length - 1 {
+		if !isASCIIDigit(isbn[i]) {
+			return false
+		}
+		weight := 1
+		if i%2 != 0 {
+			weight = 3
+		}
+		sum += int(isbn[i]-'0') * weight
+	}
+	if !isASCIIDigit(isbn[isbn13Length-1]) {
+		return false
+	}
+	return (10-sum%10)%10 == int(isbn[isbn13Length-1]-'0')
+}
+
+func referenceNormalizeISBN(s string) (string, bool) {
+	if s == "" {
+		return "", false
+	}
+
+	var builder strings.Builder
+	previousWasSeparator := false
+	for i := range len(s) {
+		switch c := s[i]; {
+		case isASCIIDigit(c) || c == 'X' || c == 'x':
+			builder.WriteByte(c)
+			previousWasSeparator = false
+		case c == '-' || c == ' ':
+			if i == 0 || i == len(s)-1 || previousWasSeparator {
+				return "", false
+			}
+			previousWasSeparator = true
+		default:
+			return "", false
+		}
+	}
+	return builder.String(), true
+}
+
+func referenceISSN(format *regexp.Regexp, s string) bool {
+	if !format.MatchString(s) {
+		return false
+	}
+
+	issn := strings.ReplaceAll(s, "-", "")
+	sum := 0
+	for i := range 7 {
+		if !isASCIIDigit(issn[i]) {
+			return false
+		}
+		sum += int(issn[i]-'0') * (8 - i)
+	}
+
+	switch checkDigit := issn[7]; {
+	case isASCIIDigit(checkDigit):
+		sum += int(checkDigit - '0')
+	case checkDigit == 'X' || checkDigit == 'x':
+		sum += 10
+	default:
+		return false
+	}
+	return sum%11 == 0
 }
